Add connect timeout retries
As shown by preliminary baking of the code, we get a fair
number of connect timeouts -- and trying it out on a dyno
seems that this happens predictably based on the machine's
load.

The behavior only manifests as a hang of sort, and normal
connections shouldn't take more than 1 second to establish.

For those cases where we do receive a hang, we can retry that
5 times (effectively making the behavior similar to a 5 sec
connect timeout) but the effect will be different based on
empirical testing.